+++ src/main/java/ai/labs/eddi/configs/dictionary/expression/RestAction.java
@@ -38,10 +38,10 @@ public class RestAction implements IRestAction {
     }
 
     @Override
-    public List<String> readActions(String workflowId, Integer packageVersion, String filter, Integer limit) {
+    public List<String> readActions(String workflowId, Integer workflowVersion, String filter, Integer limit) {
         List<String> retActions = new LinkedList<>();
         try {
-            var workflowConfiguration = workflowStore.read(workflowId, packageVersion);
+            var workflowConfiguration = workflowStore.read(workflowId, workflowVersion);
 
             List<String> actions;
             for (var workflowStep : workflowConfiguration.getWorkflowSteps()) {