+++ src/main/java/ai/labs/eddi/configs/output/keys/IRestOutputActions.java
@@ -14,6 +14,6 @@ import java.util.List;
 public interface IRestOutputActions {
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    List<String> readOutputActions(@QueryParam("workflowId") String workflowId, @QueryParam("packageVersion") Integer packageVersion,
+    List<String> readOutputActions(@QueryParam("workflowId") String workflowId, @QueryParam("workflowVersion") Integer workflowVersion,
             @QueryParam("filter") @DefaultValue("") String filter, @QueryParam("limit") @DefaultValue("20") Integer limit);
 }