+++ src/main/java/ai/labs/eddi/engine/runtime/service/IWorkflowStoreService.java
@@ -7,7 +7,7 @@ import ai.labs.eddi.configs.descriptors.model.DocumentDescriptor;
  * @author ginccc
  */
 public interface IWorkflowStoreService {
-    WorkflowConfiguration getKnowledgeWorkflow(String workflowId, Integer packageVersion) throws ServiceException;
+    WorkflowConfiguration getKnowledgeWorkflow(String workflowId, Integer workflowVersion) throws ServiceException;
 
-    DocumentDescriptor getWorkflowDocumentDescriptor(String workflowId, Integer packageVersion) throws ServiceException;
+    DocumentDescriptor getWorkflowDocumentDescriptor(String workflowId, Integer workflowVersion) throws ServiceException;
 }